Introduce process tree
Introduce a real-time process tree updating using Tracee's events.
It exposes both processes and threads information.
It is implemented to work with multiple threads, but with only one
writing thread and the others readers.
Using this process tree, we can reduce data transmission from the
kernel and unite existing data storages doing the same logic in the
future.