feat(probes): Check LSM support in kernel in addition to prog type support
The program type compatibilty is checking whether a program will successfully load in the system.
However, with LSM programs, the attachment will not be successful without the right config.
To check properly that the program will operate as expected, added to LSM programs type
compatibility check a check against the CONFIG_LSM to make sure everything will work properly.